% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_importances.R
\name{get_local_importances}
\alias{get_local_importances}
\title{Get Local Importances from a Ranger Model}
\usage{
get_local_importances(fit_object, train_data, target_variable)
}
\arguments{
\item{fit_object}{A fitted model object from a parsnip workflow, containing the ranger engine.}

\item{train_data}{The training data used to fit the model. Should contain the target variable.}

\item{target_variable}{The name of the outcome variable in the training data (as a string).}
}
\value{
A tibble summarizing the local variable importances for each class, including the mean, median, and standard deviation of the importance scores.
}
\description{
This function extracts and summarizes the local variable importance scores
from a ranger model fitted using the parsnip package.
}
