% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_classification.R
\name{rf_classification}
\alias{rf_classification}
\title{Random Forest Classification Model Tuning}
\usage{
rf_classification(
  data,
  target_variable,
  seed,
  prop,
  strata,
  trees,
  mtry_range,
  min_n_range,
  levels,
  v_folds,
  metric,
  group = NULL
)
}
\arguments{
\item{data}{A data frame containing the dataset. It should include both predictors and the target variable.}

\item{target_variable}{A string representing the name of the target variable in the dataset.}

\item{seed}{An integer to set the random seed for reproducibility.}

\item{prop}{A numeric value between 0 and 1 indicating the proportion of data to be used for training.}

\item{strata}{A string representing the name of the stratification variable for splitting the data.}

\item{trees}{An integer indicating the number of trees to grow in the Random Forest model.}

\item{mtry_range}{A numeric vector of length two specifying the range of values for the number of variables to be randomly sampled at each split.}

\item{min_n_range}{A numeric vector of length two specifying the range of values for the minimum number of observations required at each node.}

\item{levels}{An integer indicating the number of levels in the grid for tuning parameters.}

\item{v_folds}{An integer indicating the number of folds to be used in cross-validation.}

\item{metric}{A string specifying the metric to be used for model evaluation (e.g., "accuracy").}

\item{group}{(Optional) A string representing the name of the grouping variable to use for `group_initial_split`. If not provided, `initial_split` will be used.}
}
\value{
A list containing:
\describe{
  \item{best_parameters}{A data frame of the best parameters selected based on the tuning results.}
  \item{model_metrics}{A data frame of the metrics collected from the final model evaluation.}
}
}
\description{
This function performs tuning of a Random Forest classification model on a given dataset.
It includes data preprocessing, parameter tuning, and model evaluation.
}
